backend/requirements.txt
Django==2.2.16
psycopg2-binary
djangorestframework==3.12.4
Pillow==8.3.1


asgiref==3.2.10

django-filter==2.4.0

djangorestframework-simplejwt==4.8.0
gunicorn==20.0.4
PyJWT==2.1.0
pytz==2020.1
sqlparse==0.3.1